% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fmt_effect_md}
\alias{fmt_effect_md}
\title{Format an effect from a model object in markdown}
\usage{
fmt_effect_md(
  model,
  effect,
  terms = "besp",
  digits = 2,
  statistic = "t",
  b_lab = NULL,
  ci_width = 0.95
)
}
\arguments{
\item{model}{a model object}

\item{effect}{string naming an effect from a model}

\item{terms}{a string representing the terms about the effect to extract and
format and the order to print the terms. See details below. Defaults to
\code{"besp"} for parameter estimate, standard error, statistic, \emph{p}-value.}

\item{digits}{a vector of digits to use for non-\emph{p}-value terms. Defaults to
2 for 2 decimal places of precision for all terms. This argument can be a
vector to set the digits for each term, but in this case, the digits is
still ignored for \emph{p}-values.}

\item{statistic}{symbol to use for statistic. Defaults to \emph{t}.}

\item{b_lab}{label to print in subscripts after \emph{b} for when \code{"B"} is one of
the terms.}

\item{ci_width}{width to use for confidence intervals when the term \code{"i"} is
used.}
}
\description{
Format an effect from a model object in markdown
}
\details{
Currently only effects fit by \code{\link[stats:lm]{stats::lm()}} and \code{\link[lme4:lmer]{lme4::lmer()}}.

The supported terms are:
\itemize{
\item \code{"b"} - parameter estimate (think b for \emph{beta})
\item \code{"B"} - parameter estimate with a subscript label provided by \code{b_lab}
\item \code{"e"} - standard error
\item \code{"s"} - statistic. The symbol for the statistic is set by
\code{statistic}. The default value is \code{"t"} for a \emph{t}-statistic. Example
output: \emph{t} = 1.
\item \code{"S"} - statistic as in \code{"s"} but with degrees of freedom. Example
output: \emph{t}(12) = 1.
\item \code{"i"} - confidence interval. Width is set by \code{ci_width}.
\item \code{"p"} - \emph{p}-value. The p-value is formatted by \code{\link[=fmt_p_value_md]{fmt_p_value_md()}}.
}

Degrees of freedom and \emph{p}-values for \code{lmer()} models use the
Kenwood-Rogers approximation provided by \code{\link[parameters:p_value_kenward]{parameters::p_value_kenward()}}.
This computation can take a while. The confidence-interval calculation uses
default confidence interval calculation method used by
\code{\link[broom:lme4_tidiers]{broom.mixed::tidy.merMod()}}.
}
\examples{
model <- lm(breaks ~ wool * tension, warpbreaks)

# default to: b (beta), e (error), s (statistic), p (p value)
fmt_effect_md(model, "woolB", "besp")

fmt_effect_md(model, "woolB", "Besp", b_lab = "WoolB")

fmt_effect_md(model, "woolB", "i")
}
