% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_leading_zero}
\alias{fmt_leading_zero}
\title{Format numbers to remove leading zeros}
\usage{
fmt_leading_zero(xs)
}
\arguments{
\item{xs}{a vector of numbers or a character vector representing numbers}
}
\value{
the vector with leading zeros removed. This function returns a
warning if any of the values have an absolute value greater than 1.
}
\description{
Format numbers to remove leading zeros
}
\details{
APA format says that values that are bounded between \link{-1, 1} should
not be formatted with a leading zero. Common examples would be
correlations, proportions, probabilities and p-values. Why print the digit
if it's almost never used?

Zeros are printed to match the precision of the most precise number. For
example, \code{c(0, 0.111)} becomes \code{c(.000, .111)}
}
\examples{
fmt_leading_zero(c(0, 0.111))
fmt_leading_zero(c(0.99, -0.9, -0.0))
}
