% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_minus_sign}
\alias{fmt_minus_sign}
\title{Format negative numbers with a minus sign}
\usage{
fmt_minus_sign(xs)
}
\arguments{
\item{xs}{a vector of numbers or a character vector representing numbers}
}
\value{
the vector with leading hyphens replaced with HTML minus signs
(\verb{&minus;}).
}
\description{
Format negative numbers with a minus sign
}
\details{
Negative zero \code{-0}, which might happen from aggressive rounding,
does not get a minus sign.
}
\examples{
fmt_minus_sign(c(1, .2, -1, -.2))

# Don't allow zero to be signed
fmt_minus_sign(c(-0, round(-0.001)))
}
