% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{fmt_p_value_md}
\alias{fmt_p_value_md}
\title{Format a \emph{p}-value in markdown}
\usage{
fmt_p_value_md(ps)
}
\arguments{
\item{ps}{\emph{p}-values to format}
}
\value{
a character vector of markdown formatted \emph{p}-values
}
\description{
Values less than .06 are formatted with 3 digits. Values equal to .06 or
greater are formatted with 2 digits.
}
\details{
\code{\link[scales:label_pvalue]{scales::label_pvalue()}} does the initial rounding and formatting. Then this
function strips off the leading 0 of the \emph{p} value.
}
\examples{
fmt_p_value_md(0.0912)
fmt_p_value_md(0.0512)
fmt_p_value_md(0.005)

# "p less than" notation kicks in below .001.
fmt_p_value_md(0.0005)
}
