% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printy-package.R
\docType{package}
\name{printy-package}
\alias{printy}
\alias{printy-package}
\title{printy: Helper functions for pretty-printing numbers}
\description{
This package contains helper functions for formatting numbers.
}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID})

}
\keyword{internal}
