% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skel_conf_interval}
\alias{skel_conf_interval}
\alias{skel_conf_interval_v}
\title{Plug values into a confidence interval}
\usage{
skel_conf_interval_v(xs, ys, skeleton = "[{xs}, {ys}]")

skel_conf_interval(x, skeleton = "[{x[1]}, {x[2]}]")
}
\arguments{
\item{xs}{a vector of the first elements in the intervals}

\item{ys}{a vector of the second elements in the intervals}

\item{skeleton}{glue-style format to fill. defaults to \code{"[{xs}, {ys}]"} for
\code{skel_conf_interval_v()} and \code{"[{x[1]}, {x[2]}]"} for
\code{skel_conf_interval()}.}

\item{x}{a vector of two elements to plug into the confidence interval}
}
\value{
strings representing confidence intervals
}
\description{
\code{skel_conf_interval_v()} is the vectorized function. Use it to make multiple
intervals from, say, data-frame columns. \code{skel_conf_interval()} is the
unvectorized function. Use it to make a single interval from a vector of two
numbers.
}
\details{
These functions are wrappers around calls to \code{glue::glue()}.
}
\examples{
skel_conf_interval_v(c(.1, .2), c(.3, .4))
skel_conf_interval(c(.1, .3))
}
