% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{skel_range}
\alias{skel_range}
\alias{skel_range_pair}
\title{Skeleton for a range of numbers}
\usage{
skel_range_pair(x, skeleton = "{x[1]}&ndash;{x[2]}")

skel_range(xs, ys, skeleton = "{xs}&ndash;{ys}")
}
\arguments{
\item{x}{a vector of two elements to plug into the range}

\item{skeleton}{glue-style format to fill. defaults to \code{"{xs}&ndash;{ys}"} for
\code{skel_range()} and \code{"{x[1]}&ndash;{x[2]}"} for
\code{skel_range_pair()}.}

\item{xs}{a vector of the first elements in the range}

\item{ys}{a vector of the second elements in the range}
}
\value{
strings representing ranges

strings representing ranges
}
\description{
\code{skel_range()} is a vectorized function. Use it to make multiple range from,
say, data-frame columns. \code{skel_range_pair()} is the unvectorized function.
Use it to make a single range  from a vector (pair) of two numbers.

\code{skel_range()} is a vectorized function. Use it to make multiple range from,
say, data-frame columns. \code{skel_range_pair()} is the unvectorized function.
Use it to make a single range  from a vector (pair) of two numbers.
}
\details{
These functions are wrappers around calls to \code{glue::glue()}.

These functions are wrappers around calls to \code{glue::glue()}.
}
\examples{
skel_range(c(.1, .2), c(.3, .4))
skel_range_pair(c(.1, .3))
skel_range(c(.1, .2), c(.3, .4))
skel_range_pair(c(.1, .3))
}
