% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{skel_se}
\alias{skel_se}
\alias{skel_ci}
\title{Skeletons for inline stats}
\usage{
skel_se(x, skeleton = "SE&nbsp;= {x}")

skel_ci(x, ci_width = "95", skeleton = "{ci_width}\% CI&nbsp;= {x}")
}
\arguments{
\item{skeleton}{glue-style format to fill. defaults to \code{"SE&nbsp;= {x}"} for
\code{skel_se()} and \code{"95\% CI&nbsp;= {x}"} for \code{skel_ci()}.}

\item{ci_width}{width of the confidence interval to report. Defaults to
\code{"95"}.}

\item{xs}{a vector of the values to plug into the skeleton}
}
\value{
strings with stats plugged in.
}
\description{
Skeletons for inline stats
}
