% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{skel_stat_n_value}
\alias{skel_stat_n_value}
\title{Skeleton for t-statistic-like functions}
\usage{
skel_stat_n_value(x, stat = "t", skeleton = "{stat}({x[1]})&nbsp;= {x[2]}")
}
\arguments{
\item{x}{a two-element vector where the first number is the argument to the
statistical function and the second is its value.}

\item{stat}{symbol for the statistic. defaults to \code{"t"}.}

\item{skeleton}{glue-style format to fill. defaults to
\code{"{stat}({x[1]})&nbsp;= {x[2]}"}.}
}
\value{
the formatted string
}
\description{
This skeleton handles formats like t-statistics (\code{t(df) = value}) or
correlations (\code{r(n) = value}).
}
