% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr-like.R
\name{str_replace_same_as_previous}
\alias{str_replace_same_as_previous}
\title{Replace strings that duplicate the previous string}
\usage{
str_replace_same_as_previous(string, replacement)
}
\arguments{
\item{string}{a character vector}

\item{replacement}{text to use as a replacement for duplicated values}
}
\value{
a single character vector with immediately repeating items replaced
}
\description{
The common use of this function to clean up columns in a presentation-quality
table.
}
\examples{
str_replace_same_as_previous(
  c("a", "a", "a", "b", "b", "c", "d", "d"),
  ""
)
}
