% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printy-package.R
\docType{package}
\name{printy-package}
\alias{printy}
\alias{printy-package}
\title{printy: Helper functions for pretty-printing numbers}
\description{
This package contains helper functions for formatting numbers.
}
\author{
\strong{Maintainer}: TJ Mahr \email{tristan.mahr@gmail.com}

}
\keyword{internal}
